+++ max_div/solver/_presets/default.py
@@ -6,7 +6,7 @@ import math
 
 from max_div.solver import MaxDivProblem
 from max_div.solver._duration import TargetDuration, _TargetIterationCount, _TargetTimeDuration
-from max_div.solver._scheduling import ease_in, ease_out
+from max_div.solver._parameters import ease_in, ease_out
 from max_div.solver._solver_step import OptimizationStep
 from max_div.solver._strategies import InitializationStrategy, OptimizationStrategy
 from max_div.solver._strategies._initialization._presets import InitPreset